+++ examples/cpp/qb_mps_noisy/qb_mps_noisy.cpp
@@ -30,6 +30,7 @@ int main()
   s.set_svd_cutoff(scut);
   std::map<int, double> rel_scut{{0, 1.0e-3}};
   s.set_rel_svd_cutoff(rel_scut);
+  s.set_measure_sample_sequential(false);
 
   // Set the noise model
   // Uncomment the following lines to introduce noise to the simulation.