+++ include/qb/core/session_utils.hpp
@@ -71,6 +71,8 @@ namespace qb
     double rel_svd_cutoff_tnqvm;
     /// [AER backend] Simulator type (e.g., state vec, density matrix, etc.)
     std::string aer_sim_type;
+    /// [Emulator tensor network accelerators] Measurement sampling method
+    bool measure_sample_sequential;
   };
 
 }